In models, move Response/Request/Channel to separate files. #7
