+++ models/request.go
@@ -1,0 +1,6 @@
+package models
+
+type Request struct {
+	Name string `json:"name"`
+	Id   string `json:"id"`
+}
