+++ models/response.go
@@ -1,0 +1,6 @@
+package models
+
+type Response struct {
+	Request Request
+	Error   string `json:"error"`
+}
